gix problem with lazy-pages feature
