+++ pallets/gear/src/benchmarking/mod.rs
@@ -328,9 +328,12 @@ benchmarks! {
     #[extra]
     check_all {
         syscalls_integrity::main_test::<T>();
-        tests::lazy_pages_charging::<T>();
-        tests::lazy_pages_charging_special::<T>();
-        tests::lazy_pages_gas_exceed::<T>();
+        #[cfg(feature = "lazy-pages")]
+        {
+            tests::lazy_pages::lazy_pages_charging::<T>();
+            tests::lazy_pages::lazy_pages_charging_special::<T>();
+            tests::lazy_pages::lazy_pages_gas_exceed::<T>();
+        }
     } : {}
 
     #[extra]
@@ -340,17 +343,20 @@ benchmarks! {
 
     #[extra]
     check_lazy_pages_charging {
-        tests::lazy_pages_charging::<T>();
+        #[cfg(feature = "lazy-pages")]
+        tests::lazy_pages::lazy_pages_charging::<T>();
     }: {}
 
     #[extra]
     check_lazy_pages_charging_special {
-        tests::lazy_pages_charging_special::<T>();
+        #[cfg(feature = "lazy-pages")]
+        tests::lazy_pages::lazy_pages_charging_special::<T>();
     }: {}
 
     #[extra]
     check_lazy_pages_gas_exceed {
-        tests::lazy_pages_gas_exceed::<T>();
+        #[cfg(feature = "lazy-pages")]
+        tests::lazy_pages::lazy_pages_gas_exceed::<T>();
     }: {}
 
     // This bench uses `StorageMap` as a storage, due to the fact that
